fixed warnings (#1290)
* fixed warnings


* Added cosmetic suggestions

+++ include/fastrtps/utils/fixed_size_bitmap.hpp
@@ -34,18 +34,18 @@ namespace fastrtps {
 template <class T>
 struct DiffFunction
 {
-    constexpr auto operator () (T a, T b) const 
-        -> decltype(a - b) 
-    { 
-        return a - b; 
+    constexpr auto operator () (T a, T b) const
+        -> decltype(a - b)
+    {
+        return a - b;
     }
 };
 
 /**
  * Template class to hold a range of items using a custom bitmap.
  * @tparam T      Type of the elements in the range.
- *                Should have `>=` operator and `T + uint32_t` returning T. 
- * @tparam Diff   Functor calculating the difference of two T. 
+ *                Should have `>=` operator and `T + uint32_t` returning T.
+ * @tparam Diff   Functor calculating the difference of two T.
  *                The result should be assignable to a uint32_t.
  * @tparam NBITS  Size in bits of the bitmap.
  *                Range of items will be [base, base + NBITS - 1].
@@ -54,7 +54,7 @@ struct DiffFunction
 template<class T, class Diff = DiffFunction<T>, uint32_t NBITS = 256>
 class BitmapRange
 {
-    #define NITEMS ((NBITS + 31ul) / 32ul)
+    #define NITEMS ((NBITS + 31u) / 32u)
 public:
     // Alias to improve readability.
     using bitmap_type = std::array<uint32_t, NITEMS>;
@@ -74,7 +74,7 @@ public:
     /**
      * Base-specific constructor.
      * Constructs an empty range with specified base.
-     * 
+     *
      * @param base   Specific base value for the created range.
      */
     explicit BitmapRange(
@@ -92,7 +92,7 @@ public:
      * Get base of the range.
      * @return a copy of the range base.
      */
-    T base() const noexcept 
+    T base() const noexcept
     {
         return base_;
     }
@@ -148,7 +148,7 @@ public:
 
     /**
      * Returns whether the range is empty (i.e. has all bits unset).
-     * 
+     *
      * @return true if the range is empty, false otherwise.
      */
     bool empty() const noexcept
@@ -158,7 +158,7 @@ public:
 
     /**
      * Returns the highest value set in the range.
-     * 
+     *
      * @return the highest value set in the range. If the range is empty, the result is undetermined.
      */
     T max() const noexcept
@@ -175,7 +175,7 @@ public:
     {
         // Traverse through the significant items on the bitmap
         T item = base_;
-        uint32_t n_longs = (num_bits_ + 31UL) / 32UL;
+        uint32_t n_longs = (num_bits_ + 31u) / 32u;
         for (uint32_t i = 0; i < n_longs; i++)
         {
             // Check if item has at least one bit set
@@ -188,9 +188,9 @@ public:
 #if _MSC_VER
                 unsigned long bit;
                 _BitScanReverse(&bit, bits);
-                uint32_t offset = 31UL ^ bit;
+                uint32_t offset = 31u ^ bit;
 #else
-                uint32_t offset = __builtin_clz(bits);
+                uint32_t offset = static_cast<uint32_t>(__builtin_clz(bits));
 #endif
 
                 // Found first bit set in bitmap
@@ -198,7 +198,7 @@ public:
             }
 
             // There are 32 items on each bitmap item.
-            item = item + 32UL;
+            item = item + 32u;
         }
 
         return base_;
@@ -250,8 +250,8 @@ public:
             uint32_t diff = d_func(item, base_);
             num_bits_ = std::max(diff + 1, num_bits_);
             uint32_t pos = diff >> 5;
-            diff &= 31UL;
-            bitmap_[pos] |= (1UL << (31UL - diff) );
+            diff &= 31u;
+            bitmap_[pos] |= (1u << (31u - diff) );
             return true;
         }
 
@@ -291,10 +291,10 @@ public:
         num_bits_ = std::max(num_bits_, offset + n_bits);
 
         uint32_t pos = offset >> 5;             // Item position
-        offset &= 31UL;                         // Bit position inside item
+        offset &= 31u;                          // Bit position inside item
         uint32_t mask = full_mask;              // Mask with all bits set
         mask >>= offset;                        // Remove first 'offset' bits from mask
-        uint32_t bits_in_mask = 32UL - offset;  // Take note of number of set bits in mask
+        uint32_t bits_in_mask = 32u - offset;   // Take note of number of set bits in mask
 
         // This loop enters whenever the whole mask should be added
         while (n_bits >= bits_in_mask)
@@ -330,8 +330,8 @@ public:
             Diff d_func;
             uint32_t diff = d_func(item, base_);
             uint32_t pos = diff >> 5;
-            diff &= 31UL;
-            bitmap_[pos] &= ~(1UL << (31UL - diff));
+            diff &= 31u;
+            bitmap_[pos] &= ~(1u << (31u - diff));
 
             if (item == max_value)
             {
@@ -343,7 +343,7 @@ public:
     /**
      * Gets the current value of the bitmap.
      * This method is designed to be used when performing serialization of a bitmap range.
-     * 
+     *
      * @param num_bits         Upon return, it will contain the number of significant bits in the bitmap.
      * @param bitmap           Upon return, it will contain the current value of the bitmap.
      * @param num_longs_used   Upon return, it will contain the number of valid elements on the returned bitmap.
@@ -354,14 +354,14 @@ public:
             uint32_t& num_longs_used) const noexcept
     {
         num_bits = num_bits_;
-        num_longs_used = (num_bits_ + 31UL) / 32UL;
+        num_longs_used = (num_bits_ + 31u) / 32u;
         bitmap = bitmap_;
     }
 
     /**
      * Sets the current value of the bitmap.
      * This method is designed to be used when performing deserialization of a bitmap range.
-     * 
+     *
      * @param num_bits   Number of significant bits in the input bitmap.
      * @param bitmap     Points to the begining of a uint32_t array holding the input bitmap.
      */
@@ -370,9 +370,9 @@ public:
             const uint32_t* bitmap) noexcept
     {
         num_bits_ = std::min(num_bits, NBITS);
-        uint32_t num_items = ((num_bits_ + 31UL) / 32UL);
-        uint32_t num_bytes = num_items * sizeof(uint32_t);
-        bitmap_.fill(0UL);
+        uint32_t num_items = ((num_bits_ + 31u) / 32u);
+        uint32_t num_bytes = num_items * static_cast<uint32_t>(sizeof(uint32_t));
+        bitmap_.fill(0u);
         memcpy(bitmap_.data(), bitmap, num_bytes);
         calc_maximum_bit_set(num_items, 0);
     }
@@ -389,7 +389,7 @@ public:
         T item = base_;
 
         // Traverse through the significant items on the bitmap
-        uint32_t n_longs = (num_bits_ + 31UL) / 32UL;
+        uint32_t n_longs = (num_bits_ + 31u) / 32u;
         for (uint32_t i = 0; i < n_longs; i++)
         {
             // Traverse through the bits set on the item, msb first.
@@ -403,21 +403,21 @@ public:
 #if _MSC_VER
                 unsigned long bit;
                 _BitScanReverse(&bit, bits);
-                uint32_t offset = 31UL ^ bit;
+                uint32_t offset = 31u ^ bit;
 #else
-                uint32_t offset = __builtin_clz(bits);
-                uint32_t bit = 31UL ^ offset;
+                uint32_t offset = static_cast<uint32_t>(__builtin_clz(bits));
+                uint32_t bit = 31u ^ offset;
 #endif
 
                 // Call the function for the corresponding item
                 f(item + offset);
 
                 // Clear the most significant bit
-                bits &= ~(1UL << bit);
+                bits &= ~(1u << bit);
             }
 
             // There are 32 items on each bitmap item.
-            item = item + 32UL;
+            item = item + 32u;
         }
     }
 
@@ -460,7 +460,7 @@ private:
                 // bbbbbccc bbbbbbbb cccccccc dddddddd
                 // bbbbbccc cccccddd ddddd000 dddddddd
                 // bbbbbccc cccccddd ddddd000 00000000
-                uint32_t overflow_bits = 32UL - n_bits;
+                uint32_t overflow_bits = 32u - n_bits;
                 size_t last_index = NITEMS - 1u;
                 for (size_t i = 0, n = n_items; n < last_index; i++, n++)
                 {
@@ -508,7 +508,7 @@ private:
                 // aaaaaaaa bbbbbbbb aaabbbbb bbbccccc
                 // aaaaaaaa 000aaaaa aaabbbbb bbbccccc
                 // 00000000 000aaaaa aaabbbbb bbbccccc
-                uint32_t overflow_bits = 32UL - n_bits;
+                uint32_t overflow_bits = 32u - n_bits;
                 size_t last_index = NITEMS - 1u;
                 for (size_t i = last_index, n = last_index - n_items; n > 0; i--, n--)
                 {
@@ -543,17 +543,17 @@ private:
 #if _MSC_VER
                 unsigned long bit;
                 _BitScanReverse(&bit, bits);
-                uint32_t offset = (31UL ^ bit) + 1;
+                uint32_t offset = (31u ^ bit) + 1;
 #else
-                uint32_t offset = __builtin_clz(bits) + 1;
+                uint32_t offset = static_cast<uint32_t>(__builtin_clz(bits)) + 1u;
 #endif
-                num_bits_ = (i << 5UL) + offset;
+                num_bits_ = (i << 5u) + offset;
                 break;
             }
         }
     }
-}; 
-    
+};
+
 }   // namespace fastrtps
 }   // namespace eprosima
